Update core after package refactor
